+++ gazelle/python/lifecycle.go
@@ -24,22 +24,28 @@ import (
 
 var (
 	//go:embed helper.zip
-	pythonZip []byte
-	pyzPath   string
+	helperZip  []byte
+	helperPath string
 )
 
 type LifeCycleManager struct {
 	language.BaseLifecycleManager
+	pyzFilePath string
 }
 
 func (l *LifeCycleManager) Before(ctx context.Context) {
-	pyzFile, err := os.CreateTemp("", "python_zip_")
-	if err != nil {
-		log.Fatalf("failed to write parser zip: %v", err)
-	}
-	pyzPath = pyzFile.Name()
-	if _, err := pyzFile.Write(pythonZip); err != nil {
-		log.Fatalf("cannot write %q: %v", pyzPath, err)
+	helperPath = os.Getenv("GAZELLE_PYTHON_HELPER")
+	if helperPath == "" {
+		pyzFile, err := os.CreateTemp("", "python_zip_")
+		if err != nil {
+			log.Fatalf("failed to write parser zip: %v", err)
+		}
+		defer pyzFile.Close()
+		helperPath = pyzFile.Name()
+		l.pyzFilePath = helperPath
+		if _, err := pyzFile.Write(helperZip); err != nil {
+			log.Fatalf("cannot write %q: %v", helperPath, err)
+		}
 	}
 	startParserProcess(ctx)
 	startStdModuleProcess(ctx)
@@ -51,5 +57,7 @@ func (l *LifeCycleManager) DoneGeneratingRules() {
 
 func (l *LifeCycleManager) AfterResolvingDeps(ctx context.Context) {
 	shutdownStdModuleProcess()
-	os.Remove(pyzPath)
+	if l.pyzFilePath != "" {
+		os.Remove(l.pyzFilePath)
+	}
 }