+++ gazelle/python/python_test.go
@@ -21,15 +21,19 @@ package python_test
 
 import (
 	"bytes"
+	"context"
 	"errors"
-	"github.com/bazelbuild/bazel-gazelle/testtools"
-	"github.com/bazelbuild/rules_go/go/tools/bazel"
-	"github.com/ghodss/yaml"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
+	"time"
+
+	"github.com/bazelbuild/bazel-gazelle/testtools"
+	"github.com/bazelbuild/rules_go/go/runfiles"
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
+	"github.com/ghodss/yaml"
 )
 
 const (
@@ -149,11 +153,18 @@ func testPath(t *testing.T, name string, files []bazel.RunfileEntry) {
 
 		args := []string{"-build_file_name=BUILD,BUILD.bazel"}
 
-		cmd := exec.Command(gazellePath, args...)
+		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+		t.Cleanup(cancel)
+		cmd := exec.CommandContext(ctx, gazellePath, args...)
 		var stdout, stderr bytes.Buffer
 		cmd.Stdout = &stdout
 		cmd.Stderr = &stderr
 		cmd.Dir = workspaceRoot
+		helperScript, err := runfiles.Rlocation("rules_python_gazelle_plugin/python/helper")
+		if err != nil {
+			t.Fatalf("failed to initialize Python heler: %v", err)
+		}
+		cmd.Env = append(os.Environ(), "GAZELLE_PYTHON_HELPER="+helperScript)
 		if err := cmd.Run(); err != nil {
 			var e *exec.ExitError
 			if !errors.As(err, &e) {